+++ DRN.Framework.Testing/Contexts/MethodContext.cs
@@ -6,20 +6,32 @@ namespace DRN.Framework.Testing.Contexts;
 
 public class MethodContext
 {
+    public MethodContext(MethodInfo testMethod)
+    {
+        TestMethod = testMethod;
+    }
+
+
+    public MethodInfo TestMethod { get; }
     public IReadOnlyList<object> Data { get; private set; } = null!;
     public IReadOnlyList<SubstitutePair> SubstitutePairs { get; private set; } = null!;
-    public MethodInfo TestMethod { get; private set; } = null!;
+
 
     public string GetTestFolderLocation()
     {
         var testClass = TestMethod.ReflectedType!;
         var assemblyName = testClass.Assembly.GetName().Name ?? "";
-        var relativePathToTest = testClass.Namespace!.Remove(0, assemblyName.Length).TrimStart('.').Replace('.',Path.DirectorySeparatorChar);
+        var relativePathToTest = testClass.Namespace!.Remove(0, assemblyName.Length).TrimStart('.').Replace('.', Path.DirectorySeparatorChar);
         var testFolder = Path.Combine(Path.GetDirectoryName(testClass.Assembly.Location) ?? "", relativePathToTest);
 
         return testFolder;
     }
 
+    internal void SetTestData(object[] dataRow)
+    {
+        Data = dataRow;
+        SubstitutePairs = dataRow.GetSubstitutePairs();
+    }
 
     internal void ReplaceSubstitutedInterfaces(TestContext testContext)
     {
@@ -30,14 +42,5 @@ public class MethodContext
 
             testContext.ServiceCollection.ReplaceInstance(type, implementations, ServiceLifetime.Scoped);
         }
-    }
-
-
-    internal void SetMethodInfo(MethodInfo testMethod) => TestMethod = testMethod;
-
-    internal void SetTestData(object[] data)
-    {
-        Data = data;
-        SubstitutePairs = data.GetSubstitutePairs();
     }
 }