+++ DRN.Framework.Testing/DataAttributes/DataMemberAttribute.cs
@@ -2,32 +2,25 @@ namespace DRN.Framework.Testing.DataAttributes;
 
 /// <summary>
 /// Provides a data source same approach with <see cref="MemberDataAttribute"/> and generates missing data using AutoFixture and NSubstitute.
+/// Also, if test context is added as first parameter it automatically creates an instance and provides
 ///<b>This attribute can provide Complex Types that can not be provided by DataInline attributes</b>
 /// </summary>
-public class DataMemberAutoAttribute : MemberDataAttributeBase
+public class DataMemberAttribute(string methodName, params object[] methodParams) : MemberDataAttributeBase(methodName, methodParams)
 {
-    public DataMemberAutoAttribute(string methodName, params object[] methodParams) : base(methodName, methodParams)
-    {
-    }
-
     /// <inheritdoc />
     public override IEnumerable<object[]> GetData(MethodInfo testMethod)
     {
-        var data = base.GetData(testMethod)
-            .SelectMany(values => new DataInlineAutoAttribute(values).GetData(testMethod)).ToArray();
+        var data = base.GetData(testMethod).SelectMany(values => new DataInlineAttribute(values).GetData(testMethod)).ToArray();
 
         return data;
     }
-
-
+    
     /// <inheritdoc />
     protected override object[]? ConvertDataItem(MethodInfo testMethod, object? item)
     {
-        //From MemberDataAttribute.ConvertDataItem
         if (item == null) return null;
         if (item is not object[] array)
-            throw new ArgumentException(
-                $"Property {MemberName} on {MemberType ?? testMethod.DeclaringType} yielded an item that is not an object[]");
+            throw new ArgumentException($"Property {MemberName} on {MemberType ?? testMethod.DeclaringType} yielded an item that is not an object[]");
         return array;
     }
 }