+++ DRN.Test/Tests/Testing/DataAttributes/DataInlineAutoAttributeTests.cs
@@ -3,7 +3,7 @@ namespace DRN.Test.Tests.Testing.DataAttributes;
 public class DataInlineAutoAttributeTests
 {
     [Theory]
-    [DataInlineAuto(10)]
+    [DataInline(10)]
     public void AutoInline_Should_Inline_And_Auto_Generate_Missing_Test_Data(int inline, Guid autoGenerate, IMockable mock)
     {
         inline.Should().Be(10);