+++ DRN.Test/Tests/Utils/Configurations/JsonSerializerConfigurationTests.cs
@@ -3,7 +3,7 @@ namespace DRN.Test.Tests.Utils.Configurations;
 public class JsonSerializerConfigurationTests
 {
     [Theory]
-    [DataInlineContext("testDb", "Server=127.0.0.1;Port=5432;Database=myDataBase;User Id=myUsername;Password=myPassword;")]
+    [DataInline("testDb", "Server=127.0.0.1;Port=5432;Database=myDataBase;User Id=myUsername;Password=myPassword;")]
     public void Object_Should_Be_Added_As_Configuration(TestContext context, string name, string connectionString)
     {
         var connectionStrings = new ConnectionStringsCollection();
@@ -15,7 +15,7 @@ public class JsonSerializerConfigurationTests
     }
 
     [Theory]
-    [DataInlineContext("Foo", "Zoo")]
+    [DataInline("Foo", "Zoo")]
     public void Object_Should_Override_Previous_Configuration(TestContext context, string name, string connectionString)
     {
         var connectionStrings = new ConnectionStringsCollection();